refactor: rename AlertModal to AlertContainer
- Rename AlertModal.svelte -> AlertContainer.svelte for clarity
- Update imports in AlertError, AlertSimple, AlertPluginConfirm
- Better reflects component's role as a container for alerts
